LayoutTests/vibration/cancelVibration-after-pagevisibility-changed-to-hidden-expected.txt

Tests the cancelVibration after changing the value of pageVisibility to hidden.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS internals.isVibrating() is false
PASS internals.isVibrating() is true
PASS internals.isVibrating() is false
PASS successfullyParsed is true

TEST COMPLETE
